\name{colMeans}
\docType{methods}
\alias{colMeans-methods}
\alias{colMeans,ANY-method}
\alias{colMeans,gmatrix-method}

\alias{colSums-methods}
\alias{colSums,ANY-method}
\alias{colSums,gmatrix-method}

\alias{rowMeans-methods}
\alias{rowMeans,ANY-method}
\alias{rowMeans,gmatrix-method}

\alias{rowSums-methods}
\alias{rowSums,ANY-method}
\alias{rowSums,gmatrix-method}

\title{Methods to Form Row and Column Sums and Means}
\description{
This documents the \code{colMeans}, \code{rowMeans}, \code{colSums} and \code{rowSums} methods for objects of class \code{gmatrix}.
These functions form row and column sums and means for an object of class \code{gmatrix}. This is implement very
similarly to the base package (see \code{\link{rowSums}}, but the \code{na.rm} and \code{dims} arguments are not implemented).
}
\section{Methods}{
\describe{
\item{\code{signature(x = "ANY")}}{}
\item{\code{signature(x = "gmatrix")}}{}
}}
\keyword{methods}

