\name{colnames}
\docType{methods}
\alias{colnames-methods}
\alias{colnames,ANY-method}
\alias{colnames,gmatrix-method}
\alias{colnames<--methods}
\alias{colnames<-,ANY-method}
\alias{colnames<-,gmatrix-method}

\alias{rownames-methods}
\alias{rownames,ANY-method}
\alias{rownames,gmatrix-method}
\alias{rownames<--methods}
\alias{rownames<-,ANY-method}
\alias{rownames<-,gmatrix-method}

\title{Methods to get and set the 'rownames' and 'colnames' of a matrix.}
\description{
This documents the functions \code{rownames} and \code{colnames} for objects of class \code{gmatrix}. These methods are very similar to those defined in R base (see \code{\link{rownames}} in R base).
}
\section{Methods}{

\describe{
\item{\code{signature(x = "ANY")}}{}
\item{\code{signature(x = "gmatrix")}}{}
}}
\keyword{methods}

