\name{diag}
\docType{methods}
\alias{diag-methods}
\alias{diag,ANY-method}
\alias{diag,gmatrix-method}
\alias{diag,gvector-method}
\alias{diag<--methods}
\alias{diag<-,ANY-method}
\alias{diag<-,gmatrix-method}

\title{Matrix Diagonals}
\description{
Extract or replace the diagonal of a matrix, or construct a diagonal matrix. Implemented much the same as in R base. 
}
\section{Methods}{
\describe{
\item{\code{signature(x = "ANY")}}{}
\item{\code{signature(x = "gmatrix")}}{}
}}
\keyword{methods}

