\name{dim}
\docType{methods}
\alias{dim-methods}
\alias{dim,gmatrix-method}
\alias{dim<--methods}
\alias{dim<-,gmatrix-method}

\title{Dimensions of an Object}
\description{
Retrieve or set the dimension of an object. Implemented much the same as in R base. However, any redimensioning of a matrix must result in a matrix with the same number of elements.
}
\section{Methods}{
\describe{
\item{\code{signature(x = "gmatrix")}}{}
}}
\keyword{methods}
\seealso{
\code{\link[base:base-package]{dim}}
}
