\name{dimnames}
\docType{methods}
\alias{dimnames-methods}
\alias{dimnames,gmatrix-method}
\alias{dimnames<--methods}
\alias{dimnames<-,gmatrix-method}
\title{Dimnames of an Object}
\description{
Retrieve or set the dimnames of an object. Implemented much the same as R base. 
}
\section{Methods}{
\describe{
\item{\code{signature(x = "gmatrix")}}{}
}}
\keyword{methods}
\seealso{
\code{\link[base:base-package]{dimnames}}, \code{\link{colnames}}
}

