\name{skeleton-internal}
\title{Internal skeleton objects}

\alias{.check_make_valid}
\alias{.convert_to_appropriate_class}
\alias{.exprs_AB}
\alias{.exprs_Av}
\alias{.exprs_compare_e1e2}
\alias{.exprs_compare_str}
\alias{.exprs_compare_xy}
\alias{.exprs_e1e2}
\alias{.exprs_gc_e1e2}
\alias{.exprs_gc_e2e1}
\alias{.exprs_gc12_str}
\alias{.exprs_gc21_str}
\alias{.exprs_gpu_cpu_xy}
\alias{.exprs_l_e1e2}
\alias{.exprs_l_gc_e1e2}
\alias{.exprs_l_gc_e2e1}
\alias{.exprs_l_gc12_str}
\alias{.exprs_l_gc21_str}
\alias{.exprs_l_gpu_cpu_xy}
\alias{.exprs_l_str}
\alias{.exprs_l_xy}
\alias{.exprs_sf_e1e2}
\alias{.exprs_sf_gc_e1e2}
\alias{.exprs_sf_gc_e2e1}
\alias{.exprs_sf_gc12_str}
\alias{.exprs_sf_gc21_str}
\alias{.exprs_sf_gpu_cpu_xy}
\alias{.exprs_sf_str}
\alias{.exprs_sf_xy}
\alias{.exprs_str}
\alias{.exprs_xy}
\alias{.gcolon}
\alias{.gpu_get}
\alias{.onLoad}
\alias{.onUnload}
\alias{.Rclass_to_type}
\alias{.Rclass_to_type_int}
%\alias{.resetDevice}
\alias{.seq}
\alias{.silent}
\alias{.type_name}
\alias{.type_num}
\alias{.unOps}
\alias{.strOpNonExchang}
\alias{.strOpExchang}
\alias{.nonExOps}
\alias{.exOps}
\alias{.str_unaryops}


\alias{initialize-methods}
\alias{initialize,gmatrix-method}
\alias{initialize,gvector-method}

\description{Internal skeleton objects.}
\details{These are not to be called by the user.}
\keyword{internal}