\name{gtest}
\alias{gtest}


\title{
Testing
}
\description{
This function performs a large number of tests to determine if the package is working properly. Any problems are printed to the screen. It is recomended to be run at least once after installation.
}
\usage{
gtest()
}

%\arguments{
%  \item{A}{A \code{gmatrix}, \code{gvector}, \code{matrix} or \code{vector}}
%}
\value{
TRUE
}
\author{
Nathan Morris
}
