\name{gvector}
\alias{gvector}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Create a \code{gvector} object
}
\description{
Creates a gvector object of given length and initializes all elements to 0.
}
\usage{
gvector(length, type = "d")
}
\arguments{
  \item{length}{length of the object}
  \item{type}{The type (i.e. \code{"double"}, \code{"single"}, \code{"integer"} or \code{"logical"})}
}

\value{
Returns a \code{gvector} object.
}

\author{
Nathan Morris
}

\seealso{
\code{\link{vector}}
}

