\name{ifelse}
\docType{methods}
\alias{ifelse-methods}
\alias{ifelse,ANY-method}
\alias{ifelse,gmatrix-method}
\alias{ifelse,gvector-method}
\title{Conditional Element Selection}
\description{
\code{ifelse} returns a value with the same shape as test which is filled with elements selected from either \code{yes} or \code{no} depending on whether the element of test is \code{TRUE} or \code{FALSE}. 
}
\section{Methods}{
\describe{
\item{\code{signature(test = "ANY")}}{}
\item{\code{signature(test = "gmatrix")}}{}
\item{\code{signature(test = "gvector")}}{}
}}
\keyword{methods}

