\name{length}
\docType{methods}
\alias{length-methods}
\alias{length,gmatrix-method}
\alias{length,gvector-method}

\title{Length of an Object}
\description{
Get the length of a \code{gvector} or \code{gmatrix}. 
}
\section{Methods}{
\describe{
\item{\code{signature(x = "gmatrix")}}{
returns total number of elements in the matrix.
}
\item{\code{signature(x = "gvector")}}{
returns the length of the \code{gvector}
}
}}
\keyword{methods}

