\name{max}
\docType{methods}
\alias{max-methods}
\alias{max,gmatrix-method}
\alias{max,gvector-method}
\alias{min-methods}
\alias{min,gmatrix-method}
\alias{min,gvector-method}
\title{Maxima and Minima}
\description{
Returns the maxima and minima of the input values. The \code{na.rm=TRUE} option has NOT been implemented. 
}
\section{Methods}{
\describe{
\item{\code{signature(x = "gmatrix")}}{}
\item{\code{signature(x = "gvector")}}{}
}}
\keyword{methods}

