\name{mean, sum}
\docType{methods}
\alias{mean-methods}
\alias{mean,ANY-method}
\alias{mean,gmatrix-method}
\alias{mean,gvector-method}

\alias{sum-methods}
\alias{sum,gmatrix-method}
\alias{sum,gvector-method}

\title{mean and sum methods}
\description{
These methods are implement very similarly to the corresponding R base methods.
}
\section{Methods}{
\describe{
\item{\code{signature(x = "ANY")}}{}
\item{\code{signature(x = "gmatrix")}}{}
\item{\code{signature(x = "gvector")}}{}
}}
\keyword{methods}

