\name{matrix multiplication}
\docType{methods}
\alias{\%*\%-methods}
\alias{\%*\%,gmatrix,gmatrix-method}
\alias{\%*\%,gmatrix,gvector-method}
\alias{\%*\%,gmatrix,logical-method}
\alias{\%*\%,gmatrix,matrix-method}
\alias{\%*\%,gmatrix,numeric-method}
\alias{\%*\%,gvector,gmatrix-method}
\alias{\%*\%,gvector,gvector-method}
\alias{\%*\%,gvector,logical-method}
\alias{\%*\%,gvector,matrix-method}
\alias{\%*\%,gvector,numeric-method}
\alias{\%*\%,logical,gmatrix-method}
\alias{\%*\%,logical,gvector-method}
\alias{\%*\%,matrix,gmatrix-method}
\alias{\%*\%,matrix,gvector-method}
\alias{\%*\%,numeric,gmatrix-method}
\alias{\%*\%,numeric,gvector-method}
\title{Matrix Multiplication}
\description{
Multiply two matrices together.
}
\section{Methods}{
\describe{
\item{\code{signature(x = "gmatrix", y = "gmatrix")}}{}
\item{\code{signature(x = "gmatrix", y = "gvector")}}{}
\item{\code{signature(x = "gmatrix", y = "logical")}}{}
\item{\code{signature(x = "gmatrix", y = "matrix")}}{}
\item{\code{signature(x = "gmatrix", y = "numeric")}}{}
\item{\code{signature(x = "gvector", y = "gmatrix")}}{}
\item{\code{signature(x = "gvector", y = "gvector")}}{}
\item{\code{signature(x = "gvector", y = "logical")}}{}
\item{\code{signature(x = "gvector", y = "matrix")}}{}
\item{\code{signature(x = "gvector", y = "numeric")}}{}
\item{\code{signature(x = "logical", y = "gmatrix")}}{}
\item{\code{signature(x = "logical", y = "gvector")}}{}
\item{\code{signature(x = "matrix", y = "gmatrix")}}{}
\item{\code{signature(x = "matrix", y = "gvector")}}{}
\item{\code{signature(x = "numeric", y = "gmatrix")}}{}
\item{\code{signature(x = "numeric", y = "gvector")}}{}
}}
\keyword{methods}

