\name{names}
\docType{methods}
\alias{names-methods}
\alias{names,gvector-method}
\alias{names<--methods}
\alias{names<-,gvector-method}
\title{Get and set names}
\description{
Gets and sets the names slot of a \code{gvector} object.
}
\section{Methods}{
\describe{
\item{\code{signature(x = "gvector")}}{}
}}
\keyword{methods}