\name{ncol}
\docType{methods}
\alias{ncol-methods}
\alias{ncol,ANY-method}
\alias{ncol,gmatrix-method}
\alias{nrow-methods}
\alias{nrow,ANY-method}
\alias{nrow,gmatrix-method}

\title{Mumber of rows/columns}
\description{
Get the number of rows/columns of a 'gmatrix.'
}
\section{Methods}{
\describe{
\item{\code{signature(x = "ANY")}}{}

\item{\code{signature(x = "gmatrix")}}{}
}}
\keyword{methods}

