\name{show}
\docType{methods}
\alias{show-methods}
\alias{show,gmatrix-method}
\alias{show,gvector-method}
\title{Methods for displaying GPU objects}
\description{
These methods may be used to display objects of class \code{gmatrix} and 'gvector.'
}
\section{Methods}{
\describe{
\item{\code{signature(object = "gmatrix")}}{}
\item{\code{signature(object = "gvector")}}{}
}}
\keyword{methods}

