\name{sort}
\docType{methods}
\alias{sort-methods}
\alias{sort,ANY-method}
\alias{sort,gvector-method}
\alias{sort,jarrayRef-method}
\alias{sort,jrectRef-method}
\title{Sorting}
\description{
Sorts a vector
}
\section{Methods}{
\describe{
\item{\code{signature(x = "ANY")}}{}
\item{\code{signature(x = "gvector")}}{}
}}
\keyword{methods}

