\name{t}
\docType{methods}
\alias{t-methods}
\alias{t,ANY-method}
\alias{t,gmatrix-method}
\alias{t,gvector-method}
\alias{t.gmatrix}
\alias{t.gvector}
\title{Transpose}
\description{
Methods to traspose a \code{gmatrix} or \code{gvector}.
}
\section{Methods}{
\describe{
\item{\code{signature(x = "ANY")}}{}
\item{\code{signature(x = "gmatrix")}}{}
\item{\code{signature(x = "gvector")}}{}
}}
\keyword{methods}

