\name{which}
\docType{methods}
\alias{which-methods}
\alias{which,ANY-method}
\alias{which,gvector-method}
\title{Which indices are TRUE?}
\description{
Give the TRUE indices of a logical \code{gvector}. 
}
\section{Methods}{
\describe{

\item{\code{signature(x = "ANY")}}{}

\item{\code{signature(x = "gvector")}}{}
}}
\keyword{methods}